+++ java-chess/src/main/java/com/sarunasbend/github/logic/Chessboard.java
@@ -23,6 +23,7 @@ public class Chessboard {
         Graphics g = this.chessboardImage.createGraphics();
         Graphics2D g2d = (Graphics2D) g;
 
+        // anti-aliasing looks kinda bad with the smaller font
         // g2d.setRenderingHint(RenderingHints.KEY_TEXT_ANTIALIASING,RenderingHints.VALUE_TEXT_ANTIALIAS_ON);
 
         boolean isWhite = true;
