--- java-chess/src/main/java/com/sarunasbend/github/logic/pieces/Bishop.java
@@ -1,7 +1,0 @@
-package com.sarunasbend.github.logic.pieces;
-
-public class Bishop extends Piece {
-    public Bishop(String id){
-        super(id, "Bishop");
-    }    
-}
