+++ java-chess/src/main/java/com/sarunasbend/github/logic/pieces/BlackPieces.java
@@ -1,0 +1,7 @@
+package com.sarunasbend.github.logic.pieces;
+
+public class BlackPieces {
+    public BlackPieces(){
+        
+    }
+}
