--- java-chess/src/main/java/com/sarunasbend/github/logic/pieces/King.java
@@ -1,7 +1,0 @@
-package com.sarunasbend.github.logic.pieces;
-
-public class King extends Piece {
-   public King(String id){
-        super(id, "King");
-   }
-}
