--- java-chess/src/main/java/com/sarunasbend/github/logic/pieces/Knight.java
@@ -1,7 +1,0 @@
-package com.sarunasbend.github.logic.pieces;
-
-public class Knight extends Piece {
-    public Knight(String id){
-        super(id, "Knight");
-    }
-}
