--- java-chess/src/main/java/com/sarunasbend/github/logic/pieces/Pawn.java
@@ -1,8 +1,0 @@
-package com.sarunasbend.github.logic.pieces;
-
-public class Pawn extends Piece{
-    public Pawn(String id){
-        super(id, "Pawn");
-    }
-    
-}
