--- java-chess/src/main/java/com/sarunasbend/github/logic/pieces/Queen.java
@@ -1,7 +1,0 @@
-package com.sarunasbend.github.logic.pieces;
-
-public class Queen extends Piece {
-    public Queen(String id){
-        super(id, "Queen");
-    }
-}
