--- java-chess/src/main/java/com/sarunasbend/github/logic/pieces/Rook.java
@@ -1,7 +1,0 @@
-package com.sarunasbend.github.logic.pieces;
-
-public class Rook extends Piece{
-    public Rook(String id){
-        super(id, "Rook");
-    }    
-}
