+++ java-chess/src/main/java/com/sarunasbend/github/logic/pieces/WhitePieces.java
@@ -1,0 +1,5 @@
+package com.sarunasbend.github.logic.pieces;
+
+public class WhitePieces {
+    
+}
