+++ java-chess/src/main/java/com/sarunasbend/github/logic/pieces/piece/Bishop.java
@@ -1,0 +1,8 @@
+package com.sarunasbend.github.logic.pieces.piece;
+
+public class Bishop extends Piece {
+    public Bishop(String id, String name){
+        super(id, name);
+    }
+    
+}
