+++ java-chess/src/main/java/com/sarunasbend/github/logic/pieces/piece/King.java
@@ -1,0 +1,7 @@
+package com.sarunasbend.github.logic.pieces.piece;
+
+public class King extends Piece {
+   public King(String id, String name){
+        super(id, name);
+   }
+}
