+++ java-chess/src/main/java/com/sarunasbend/github/logic/pieces/piece/Knight.java
@@ -1,0 +1,7 @@
+package com.sarunasbend.github.logic.pieces.piece;
+
+public class Knight extends Piece {
+    public Knight(String id, String name){
+        super(id, name);
+    }
+}
