+++ java-chess/src/main/java/com/sarunasbend/github/logic/pieces/piece/Pawn.java
@@ -1,0 +1,7 @@
+package com.sarunasbend.github.logic.pieces.piece;
+
+public class Pawn extends Piece{
+    public Pawn(String id, String name){
+        super(id, name);
+    }   
+}
