+++ java-chess/src/main/java/com/sarunasbend/github/logic/pieces/piece/Queen.java
@@ -1,0 +1,7 @@
+package com.sarunasbend.github.logic.pieces.piece;
+
+public class Queen extends Piece {
+    public Queen(String id, String name){
+        super(id, name);
+    }
+}
