+++ java-chess/src/main/java/com/sarunasbend/github/logic/pieces/piece/Rook.java
@@ -1,0 +1,7 @@
+package com.sarunasbend.github.logic.pieces.piece;
+
+public class Rook extends Piece{
+    public Rook(String id, String name){
+        super(id, name);
+    }    
+}
